Move files to src dir
