+++ lib/api.dart
@@ -1,8 +1,8 @@
 library aspose_barcode_cloud.api;
 
-export './configuration.dart' show Configuration;
-export 'api_client.dart' show ApiClient;
-export 'api_exception.dart' show ApiException;
+export 'src/configuration.dart' show Configuration;
+export 'src/api_client.dart' show ApiClient;
+export 'src/api_exception.dart' show ApiException;
 
 export 'api/barcode_api.dart' show BarcodeApi;
 export 'api/file_api.dart' show FileApi;
