+++ lib/api/file_api.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data' show Uint8List;
 import 'package:http/http.dart' show MultipartFile, MultipartRequest;
 
 import '../api.dart';
-import '../api_helper.dart';
+import '../src/api_helper.dart';
 
 class FileApi {
   FileApi(this.apiClient) {}
