+++ lib/src/api_exception.dart
