+++ lib/src/api_helper.dart
@@ -1,4 +1,4 @@
-import 'api.dart';
+import '../api.dart';
 
 const _delimiters = const {'csv': ',', 'ssv': ' ', 'tsv': '\t', 'pipes': '|'};
